Make curves plugin accept three control points
Previously it accepted just four. You could of course duplicate the middle
ones. It's neater to deal with it this way, though.

In case the plugin is provided a wrong amount of control points, it will just
return instead of fail. Might want to change this behavior somehow (give a
warning or so).

Related to #36.